Revert "Add role-based access control for route protection"
